task4

+++ 0x06-pointers_arrays_strings/4-rev_array.c
@@ -8,10 +8,10 @@
  */
 void reverse_array(int *a, int n)
 {
-	int i;
+	long int i;
 
 	for (i = (n - 1); i > -1; i--)
 	{
-		printf("%d", a[i]);
+		printf("%ld", a[i]);
 	}
 }
